Move new visitors to internal package
