+++ pmd-java/src/main/java/net/sourceforge/pmd/lang/java/metrics/impl/NcssMetric.java
@@ -8,7 +8,7 @@ import org.apache.commons.lang3.mutable.MutableInt;
 
 import net.sourceforge.pmd.lang.java.ast.ASTAnyTypeDeclaration;
 import net.sourceforge.pmd.lang.java.ast.MethodLikeNode;
-import net.sourceforge.pmd.lang.java.metrics.impl.visitors.NcssVisitor;
+import net.sourceforge.pmd.lang.java.metrics.impl.internal.NcssVisitor;
 import net.sourceforge.pmd.lang.metrics.MetricOption;
 import net.sourceforge.pmd.lang.metrics.MetricOptions;
